Move Verilator tests to new folder and update names
