+++ tests/verilator/Makefile
@@ -16,7 +16,7 @@ endif
 
 
 default:
-	$(VERILATOR) -f vargs.vc --trace-fst -cc --exe --build --trace rvsteel_sim_verilator.v vcfg.vlt main.cpp argparse.cpp -o rvsteel_sim_verilator
+	$(VERILATOR) -f vargs.vc --trace-fst -cc --exe --build --trace unit_tests.v vcfg.vlt main.cpp argparse.cpp -o unit_tests
